Extract branch name warning UI
